Separated base classes from core into new file. Created new file for tooth curves. Renamed core InvoluteGear to just Gear - preparing for other types. Updated wrapper to keep things working.
