+++ src/gggears/gggears_build123d.py
@@ -35,7 +35,7 @@ class GearBuilder(GearToNurbs):
 
     def __init__(
         self,
-        gear: gg.InvoluteGear,
+        gear: gg.Gear,
         n_points_hz=4,
         n_points_vert=4,
         oversampling_ratio=2.5,
